Replace DSpaceConfigurationFactoryBean with ConfigurationPropertiesFactoryBean
